refactor: sync/init command
