+++ internal/pkg/service/cli/cmd/sync/init/dialog.go
@@ -2,10 +2,16 @@ package init
 
 import (
 	"context"
+	"fmt"
+	"math"
+	"strings"
 
 	"github.com/keboola/go-client/pkg/keboola"
+	"github.com/keboola/go-utils/pkg/orderedmap"
 
+	"github.com/keboola/keboola-as-code/internal/pkg/model"
 	"github.com/keboola/keboola-as-code/internal/pkg/naming"
+	"github.com/keboola/keboola-as-code/internal/pkg/search"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/cli/cmd/ci/workflow"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/cli/dialog"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/cli/prompt"
@@ -13,6 +19,13 @@ import (
 	createManifest "github.com/keboola/keboola-as-code/pkg/lib/operation/project/local/manifest/create"
 	workflowsGen "github.com/keboola/keboola-as-code/pkg/lib/operation/project/local/workflows/generate"
 	initOp "github.com/keboola/keboola-as-code/pkg/lib/operation/project/sync/init"
+)
+
+const (
+	ModeMainBranch     = "only main branch"
+	ModeAllBranches    = "all branches"
+	ModeSelectSpecific = "select branches"
+	ModeTypeList       = "type IDs or names"
 )
 
 type initDeps interface {
@@ -29,7 +42,7 @@ func AskInitOptions(ctx context.Context, d *dialog.Dialogs, dep initDeps, f Flag
 	}
 
 	// Allowed branches
-	if allowedBranches, err := d.AskAllowedBranches(ctx, dep); err == nil {
+	if allowedBranches, err := AskAllowedBranches(ctx, dep, d, f); err == nil {
 		out.ManifestOptions.AllowedBranches = allowedBranches
 	} else {
 		return out, err
@@ -58,4 +71,203 @@ func AskInitOptions(ctx context.Context, d *dialog.Dialogs, dep initDeps, f Flag
 	}
 
 	return out, nil
+}
+
+type branchesDialog struct {
+	*dialog.Dialogs
+	Flags
+	deps        branchesDialogDeps
+	allBranches []*model.Branch
+}
+
+type branchesDialogDeps interface {
+	KeboolaProjectAPI() *keboola.AuthorizedAPI
+}
+
+func AskAllowedBranches(ctx context.Context, deps branchesDialogDeps, d *dialog.Dialogs, f Flags) (model.AllowedBranches, error) {
+	return (&branchesDialog{Dialogs: d, deps: deps, Flags: f}).ask(ctx)
+}
+
+func (d *branchesDialog) ask(ctx context.Context) (model.AllowedBranches, error) {
+	// Get Storage API
+	api := d.deps.KeboolaProjectAPI()
+
+	// List all branches
+	if v, err := api.ListBranchesRequest().Send(ctx); err == nil {
+		for _, apiBranch := range *v {
+			d.allBranches = append(d.allBranches, model.NewBranch(apiBranch))
+		}
+	} else {
+		return nil, err
+	}
+
+	// Defined by flag
+	if d.Branches.IsSet() {
+		value := d.Branches.Value
+		if value == "*" {
+			return model.AllowedBranches{model.AllBranchesDef}, nil
+		} else if value == "main" {
+			return model.AllowedBranches{model.MainBranchDef}, nil
+		}
+		if allowedBranches := d.parseBranchesList(value, `,`); len(allowedBranches) > 0 {
+			return allowedBranches, nil
+		}
+		return nil, errors.New(`please specify at least one branch`)
+	}
+
+	// Ask user
+	switch d.askMode() {
+	case ModeMainBranch:
+		return model.AllowedBranches{model.MainBranchDef}, nil
+	case ModeAllBranches:
+		return model.AllowedBranches{model.AllBranchesDef}, nil
+	case ModeSelectSpecific:
+		if selectedBranches, err := SelectBranches(d.allBranches, `Allowed project's branches:`, d.Dialogs, d.Flags); err == nil {
+			return branchesToAllowedBranches(selectedBranches), nil
+		} else {
+			return nil, err
+		}
+	case ModeTypeList:
+		if results := d.askBranchesList(); len(results) > 0 {
+			return results, nil
+		}
+	}
+
+	return nil, errors.New(`please specify at least one branch`)
+}
+
+func (d *branchesDialog) askMode() string {
+	mode, _ := d.Select(&prompt.Select{
+		Label: "Allowed project's branches:",
+		Description: "Please select which project's branches you want to use with this CLI.\n" +
+			"The other branches will still exist, but they will be invisible in the CLI.",
+		Options: []string{
+			ModeMainBranch,
+			ModeAllBranches,
+			ModeSelectSpecific,
+			ModeTypeList,
+		},
+		Default: ModeMainBranch,
+	})
+	return mode
+}
+
+func (d *branchesDialog) askBranchesList() model.AllowedBranches {
+	// Print first 10 branches for inspiration
+	end := math.Min(10, float64(len(d.allBranches)))
+	d.Printf("\nExisting project's branches, for inspiration:\n")
+	for _, branch := range d.allBranches[:int(end)] {
+		d.Printf("%s (%d)\n", branch.Name, branch.ID)
+	}
+	if len(d.allBranches) > 10 {
+		d.Printf(`...`)
+	}
+
+	// Prompt
+	lines, ok := d.Multiline(&prompt.Question{
+		Label: "Allowed project's branches:",
+		Description: "\nPlease enter one branch definition per line.\n" +
+			"Each definition can be:\n" +
+			"- branch ID\n" +
+			"- branch name, with optional wildcards, eg. \"Foo Bar\", \"Dev:*\"\n" +
+			"- branch directory (normalized) name, with optional wildcards, eg. \"foo-bar\", \"dev-*\"\n",
+		Validator: func(val any) error {
+			// At least one existing branch must match user definition
+			matched := 0
+			for _, branch := range d.allBranches {
+				for _, definition := range d.parseBranchesList(val.(string), "\n") {
+					if definition.IsBranchAllowed(branch) {
+						matched++
+					}
+				}
+			}
+			if matched == 0 {
+				return errors.New(`no existing project's branch matches your definitions`)
+			}
+			return nil
+		},
+	})
+
+	if !ok {
+		return nil
+	}
+
+	// Normalize
+	return d.parseBranchesList(lines, "\n")
+}
+
+func (d *branchesDialog) parseBranchesList(str, sep string) model.AllowedBranches {
+	branches := model.AllowedBranches{}
+	for _, item := range strings.Split(str, sep) {
+		item = strings.TrimSpace(item)
+		if len(item) == 0 {
+			continue
+		}
+		branches = append(branches, model.AllowedBranch(item))
+	}
+	return d.unique(branches)
+}
+
+// unique returns only unique items.
+func (d *branchesDialog) unique(items model.AllowedBranches) model.AllowedBranches {
+	m := orderedmap.New()
+	for _, item := range items {
+		m.Set(string(item), true)
+	}
+
+	unique := model.AllowedBranches{}
+	for _, item := range m.Keys() {
+		unique = append(unique, model.AllowedBranch(item))
+	}
+	return unique
+}
+
+func branchesToAllowedBranches(branches []*model.Branch) (out model.AllowedBranches) {
+	for _, b := range branches {
+		out = append(out, model.AllowedBranch(b.ID.String()))
+	}
+	return out
+}
+
+func SelectBranches(all []*model.Branch, label string, d *dialog.Dialogs, f Flags) (results []*model.Branch, err error) {
+	if f.Branches.IsSet() {
+		errs := errors.NewMultiError()
+		for _, item := range strings.Split(f.Branches.Value, `,`) {
+			item = strings.TrimSpace(item)
+			if len(item) == 0 {
+				continue
+			}
+
+			if b, err := search.Branch(all, item); err == nil {
+				results = append(results, b)
+			} else {
+				errs.Append(err)
+				continue
+			}
+		}
+		if len(results) > 0 {
+			return results, errs.ErrorOrNil()
+		}
+		return nil, errors.New(`please specify at least one branch`)
+	}
+
+	selectOpts := orderedmap.New()
+	for _, branch := range all {
+		msg := fmt.Sprintf(`%s (%d)`, branch.Name, branch.ID)
+		selectOpts.Set(msg, branch.ID)
+	}
+	indexes, _ := d.MultiSelectIndex(&prompt.MultiSelectIndex{
+		Label:       label,
+		Description: "Please select one or more branches.",
+		Options:     selectOpts.Keys(),
+		Validator:   prompt.AtLeastOneRequired,
+	})
+	for _, index := range indexes {
+		results = append(results, all[index])
+	}
+	if len(results) > 0 {
+		return results, nil
+	}
+
+	return nil, errors.New(`please specify at least one branch`)
 }
